Set outputs for networking module
"Notice that the VPC output passes a reference to the entire output of the VPC module.
This allows us to be succinct in the outputs code, especially when passing data through
multiple layers of nested modules. Also notice that the sg output is made up of a new
object containing the ids of the security groups. This pattern is useful for grouping
related attributes from different resources together into a single output value."

Source: Terraform in Action by Scott Winkler